✅ Update fetch_schemas to accept optional fields
The "fax" and "type" fields in the fetch_schemas.py file have been changed to optional. This modification allows for instances where these details may not be readily available or necessary for every location. 'Required=False' for these fields have been replaced with 'default=None' setting a default value of None when the data is not provided.